лаб 6,7

+++ lab3.2.py
@@ -1,5 +1,5 @@
 import sys
 bar = sys.argv[1]
-foo = ((int(bar) ** 2)/4)/3.14
+foo = ((float(bar) ** 2)/4)/3.14
 sys.stdout.write('площа кола = ')
 sys.stdout.write(str(foo))
